fix image build failure (#335)
# Remove the platform parameter from Dockerfile
We should remove the fixed `--platform=linux/amd64` from Dockefile to
allow docker to choose right architecture when pulling the image.

If this change affects the github action, we should specify the platform
parameter on github action task

# Fix poetry version issue
When I try to build the image using `docker compose build` today, it
failed due to a breaking change of poetry 2.0.0.

We need to put a constrain of 'poetry<2.0.0' for now.

```
docker compose build

> [django-socio-grpc server 1/1] RUN poetry install:
0.493
0.502 The Poetry configuration is invalid:
0.502   - project must contain ['name'] properties
```

Others are experiencing [the same
issue](https://github.com/HumanSignal/label-studio/issues/6851)

Co-authored-by: Randall Wang <randall@talpa-solutions.com>